The Dzaleka Art Project
The Dzaleka Art Project is a community-based collaborative project by and about the arts and artists living as refugees in the Dzaleka camp in Malawi. Enjoy the talent and creativity in the camp while getting to know the dreams and struggles of some of its residents. Please consider visiting the website that complements the book and share what you learn with others.